DAOS-10922 control: Implement ipmctl stubs for non-amd64 platforms
Implement NvmMgmt and IpmCtl interface for platforms that don't support ipmctl.
